fix: better snow fall effect
