+++ packages/nc-gui/components/smartsheet/Topbar.vue
@@ -1,4 +1,7 @@
 <script lang="ts" setup>
+import dayjs from 'dayjs'
+import { useStorage } from '@vueuse/core'
+
 const { isGrid, isForm, isGallery, isKanban, isMap, isCalendar } = useSmartsheetStoreOrThrow()
 
 const router = useRouter()
@@ -27,6 +30,19 @@ const topbarBreadcrumbItemWidth = computed(() => {
     return 'calc(\(100% - 12px\) / 2)'
   }
 })
+
+const isSnowFallEnabled = useStorage(
+  EasterEggs.SNOWFLAKE_ENABLED,
+  dayjs().isBetween(dayjs(`${dayjs().year()}-12-20`), dayjs(`${dayjs().year()}-12-28`)),
+)
+
+const shouldShowToggleMenu = computed(() => {
+  return isSnowFallEnabled.value || dayjs().isBetween(dayjs(`${dayjs().year()}-12-20`), dayjs(`${dayjs().year()}-12-31`))
+})
+
+const toggleSnowFall = () => {
+  isSnowFallEnabled.value = !isSnowFallEnabled.value
+}
 </script>
 
 <template>
@@ -55,6 +71,24 @@ const topbarBreadcrumbItemWidth = computed(() => {
       <div class="flex items-center justify-end gap-2 flex-1">
         <GeneralApiLoader v-if="!isMobileMode" />
 
+        <NcButton
+          v-if="shouldShowToggleMenu"
+          v-e="['c:easter-egg-snowfall']"
+          type="secondary"
+          size="small"
+          class="nc-topbar-snowfall-btn"
+          :class="{ '!bg-brand-50 !hover:bg-brand-100/70 !text-brand-500': isSnowFallEnabled }"
+          data-testid="nc-topbar-snowfall-btn"
+          @click="toggleSnowFall"
+        >
+          <div class="flex items-center justify-center min-w-[28.69px]">
+            <GeneralIcon
+              icon="snow"
+              class="w-4 h-4 !stroke-transparent"
+              :class="{ 'border-l-1 border-transparent': isSnowFallEnabled }"
+            />
+          </div>
+        </NcButton>
         <NcButton
           v-if="!isSharedBase && isFeatureEnabled(FEATURE_FLAG.EXTENSIONS) && openedViewsTab === 'view'"
           v-e="['c:extension-toggle']"
